updated finder.sh to take in arguments 1 and 2 see if the first argument is created, then check the number of files and check the matching lines within those files. then echo's the number of files and nummatchs
Updated writer.sh to take in arguments 1 and 2, check if there are more than 2 arguements, if so exit 1. Creates the direction of arguement 1, if the director is not created, will have an exit 1. Then echos the second argument and pushs that string to the file created in arugment 1. if that operation is successful will have an exit 0, if not will have an exit 1.

both ./finder-test.sh and ./full-test.sh successfully ran and completed without errors.